# docs/api/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "logaffe",
    "description": "The read and write surface of one logaffe installation. Every installation is at its own address.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/projects/{projectId}/ingest-tokens": {
      "post": {
        "tags": [
          "logaffe"
        ],
        "summary": "Gives a project a token to receive on.",
        "operationId": "IssueIngestToken",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssuedIngestTokenResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "409": {
            "description": "Conflict"
          }
        }
      },
      "get": {
        "tags": [
          "logaffe"
        ],
        "summary": "What one project can currently receive on.",
        "operationId": "ListIngestTokens",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ListedIngestTokenResponse"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/ingest-tokens/{id}/token": {
      "get": {
        "tags": [
          "logaffe"
        ],
        "summary": "The token that is in the row, and the delivery to paste it in.",
        "operationId": "ReadIngestTokenBack",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadIngestTokenResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/ingest-tokens/{id}": {
      "delete": {
        "tags": [
          "logaffe"
        ],
        "summary": "Ends a token, immediately.",
        "operationId": "RevokeIngestToken",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/agent-tokens": {
      "post": {
        "tags": [
          "logaffe"
        ],
        "summary": "Gives an agent a token to read with.",
        "operationId": "IssueAgentToken",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IssueAgentTokenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssuedAgentTokenResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "logaffe"
        ],
        "summary": "Every agent token the installation holds.",
        "operationId": "ListAgentTokens",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ListedAgentTokenResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/agent-tokens/{id}": {
      "patch": {
        "tags": [
          "logaffe"
        ],
        "summary": "Gives an agent token another label.",
        "operationId": "RenameAgentToken",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RenameAgentTokenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "logaffe"
        ],
        "summary": "Ends a token, immediately.",
        "operationId": "RevokeAgentToken",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/agent-tokens/{id}/token": {
      "get": {
        "tags": [
          "logaffe"
        ],
        "summary": "The token that is in the row, and the configuration to paste it in.",
        "operationId": "ReadAgentTokenBack",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadAgentTokenResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/hosts/{hostId}/host-tokens": {
      "post": {
        "tags": [
          "logaffe"
        ],
        "summary": "Gives a host a token its collector can report on.",
        "operationId": "IssueHostToken",
        "parameters": [
          {
            "name": "hostId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssuedHostTokenResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "409": {
            "description": "Conflict"
          }
        }
      },
      "get": {
        "tags": [
          "logaffe"
        ],
        "summary": "What one host can currently receive samples on.",
        "operationId": "ListHostTokens",
        "parameters": [
          {
            "name": "hostId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ListedHostTokenResponse"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/host-tokens/{id}/token": {
      "get": {
        "tags": [
          "logaffe"
        ],
        "summary": "The token that is in the row, and the command to paste it in.",
        "operationId": "ReadHostTokenBack",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadHostTokenResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/host-tokens/{id}": {
      "delete": {
        "tags": [
          "logaffe"
        ],
        "summary": "Ends a token, immediately.",
        "operationId": "RevokeHostToken",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "HttpValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      },
      "IssueAgentTokenRequest": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "IssuedAgentTokenResponse": {
        "required": [
          "id",
          "name",
          "token",
          "clientConfiguration",
          "issuedAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "token": {
            "type": "string"
          },
          "clientConfiguration": {
            "type": "string"
          },
          "issuedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "IssuedIngestTokenResponse": {
        "required": [
          "id",
          "token",
          "deliverySnippet",
          "issuedAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "token": {
            "type": "string"
          },
          "deliverySnippet": {
            "type": "string"
          },
          "issuedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "ListedIngestTokenResponse": {
        "required": [
          "id",
          "identifier",
          "issuedAt",
          "lastUsedAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "identifier": {
            "type": "string"
          },
          "issuedAt": {
            "type": "string",
            "format": "date-time"
          },
          "lastUsedAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          }
        }
      },
...
      "ReadIngestTokenResponse": {
        "required": [
          "token",
          "deliverySnippet"
        ],
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          },
          "deliverySnippet": {
            "type": "string"
          }
        }
      },
...
      "RenameAgentTokenRequest": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
    }
  },
...
}